Pmod Bridge: Remove ports shared between interfaces
Originally, some ports shared the FPGA-side ports with other interfaces. This led to highly-confusing enablement dependency conditions. Splitting out all of the ports so that each interface has its own set makes it so that the dependency conditions are only required at the interface level instead of also at the port level.